+++ back-end/src/database/models/Sale.js
@@ -10,7 +10,7 @@ module.exports = (sequelize, DataTypes) => {
     },
     userId: {type: DataTypes.INTEGER, allowNull: false },
     sellerId: {type: DataTypes.INTEGER, allowNull: false },
-    totalPrice: {type: DataTypes.INTEGER, allowNull: false },
+    totalPrice: {type: DataTypes.DECIMAL(9, 2), allowNull: false },
     deliveryAddress:{type: DataTypes.STRING, allowNull: false },
     deliveryNumber: {type: DataTypes.STRING, allowNull: false },
     saleDate: { type: DataTypes.DATE, defaultValue: DataTypes.NOW },
